requirements.txt

requests==2.32.3
python-dotenv==1.0.1
web3==6.19.0
celery==5.4.0
aioredis==2.0.1